refactor(ui): implement adaptive flag pickers for currency and language
