+++ packages/world_countries/lib/src/widgets/currency/currency_picker.dart
@@ -1,7 +1,9 @@
 import "package:flutter/gestures.dart" show DragStartBehavior;
 import "package:flutter/material.dart";
+import "package:meta/meta.dart";
 import "package:world_flags/world_flags.dart";
 
+import "../../extensions/pickers/basic_picker_flags_extension.dart";
 import "../../extensions/world_countries_build_context_extension.dart";
 import "../../models/item_properties.dart";
 import "../../models/locale/typed_locale.dart";
@@ -58,6 +60,101 @@ class CurrencyPicker extends BasicPicker<FiatCurrency> {
     super.translation,
     super.flagsMap,
   }) : super(currencies);
+
+  /// A factory constructor that creates a [CurrencyPicker] with a pre-computed
+  /// map of flags based on the device's locale or a provided [localeCountry].
+  ///
+  /// Unlike the default constructor which uses a manually built [flagsMap],
+  /// this constructor automatically computes the most relevant flag for each
+  /// currency. It prioritizes the user's geopolitical context to provide a more
+  /// intuitive experience, saving you from complex manual mapping.
+  ///
+  /// For example, if the user's device is set to Greenland (`GL`), the Danish
+  /// Krone (`DKK`) will be displayed with the Greenlandic flag 🇬🇱. For all
+  /// other users, it will show the flag of Denmark 🇩🇰.
+  ///
+  /// This is achieved by first checking the provided [localeCountry]. If it's
+  /// `null`, it inspects the user's preferred locales from the platform. The
+  /// first locale with a valid country code is then used as the context. Flags
+  /// provided in the initial [flagsMap] are respected and will not be
+  /// overridden by this logic.
+  @experimental
+  CurrencyPicker.adaptiveFlags({
+    Iterable<FiatCurrency> currencies = FiatCurrency.list,
+    super.addAutomaticKeepAlives,
+    super.addRepaintBoundaries,
+    super.addSemanticIndexes,
+    super.cacheExtent,
+    super.caseSensitiveSearch,
+    super.chosen,
+    super.clipBehavior,
+    super.crossAxisAlignment,
+    super.direction,
+    super.disabled,
+    super.dragStartBehavior,
+    super.emptyStatePlaceholder,
+    super.itemBuilder,
+    super.key,
+    super.keyboardDismissBehavior,
+    super.mainAxisAlignment,
+    super.mainAxisSize,
+    super.onSelect,
+    super.padding,
+    super.physics,
+    super.primary,
+    super.restorationId,
+    super.reverse,
+    super.scrollController,
+    super.searchBar,
+    super.searchBarPadding,
+    super.searchIn,
+    super.onSearchResultsBuilder,
+    super.separator,
+    super.showClearButton,
+    super.showSearchBar,
+    super.shrinkWrap,
+    super.sort,
+    super.startWithSearch,
+    super.textBaseline,
+    super.textDirection,
+    super.verticalDirection,
+    super.spacing,
+    super.translation,
+    @mustBeConst Map<FiatCurrency, BasicFlag> flagsMap = defaultFlagsMap,
+
+    /// The country to use as the locale reference for flag selection and
+    /// mapping. If provided, this country will be used to determine the default
+    /// flag for currencies that are associated with a specific country context.
+    /// If `null`, the picker will attempt to infer the locale country from the
+    /// current context. For example if device's country is Greenland (`GL`),
+    /// the picker will use the Greenlandic 🇬🇱 flag for the Danish Krone
+    /// (`DKK`), but if device's country is China (`CN`), the picker will use
+    /// the flag of the Denmark 🇩🇰 for the Danish Krone (`DKK`).
+    WorldCountry? localeCountry,
+
+    /// An optional custom function to map or transform a [BasicFlag] for a
+    /// given currency. This function allows you to override the default flag
+    /// appearance or logic for each currency and its associated countries. If
+    /// not provided, a default flag mapping will be used (simmilar to country
+    /// picker).
+    BasicFlag Function(
+      BasicFlag flag,
+      FiatCurrency currency,
+      List<WorldCountry>? countries,
+    )?
+    flagMapper,
+  }) : super(
+         currencies,
+         flagsMap: flagsMap.adaptFlags(
+           currencies.byCountryMap(),
+           flagsMapper: flagMapper,
+           localeCountry: localeCountry,
+         ),
+       );
+
+  /// Default flags map for the [CurrencyPicker.adaptiveFlags].
+  /// Mapping Euro currency to the European Union 🇪🇺 flag.
+  static const defaultFlagsMap = {FiatEur(): StarFlag(flagEurProperties)};
 
   @override
   Widget defaultBuilder(
