+++ packages/world_countries/lib/world_countries.dart
@@ -36,7 +36,6 @@ export "src/widgets/base_widgets/stateful_searchable.dart";
 export "src/widgets/country/country_picker.dart";
 export "src/widgets/country/country_tile.dart";
 export "src/widgets/country/emoji_flag.dart";
-export "src/widgets/currency/currency_flag_picker.dart";
 export "src/widgets/currency/currency_picker.dart";
 export "src/widgets/currency/currency_tile.dart";
 export "src/widgets/generic_widgets/implicit_search_delegate.dart";
@@ -45,7 +44,6 @@ export "src/widgets/generic_widgets/list_item_tile.dart";
 export "src/widgets/generic_widgets/search_list_listenable_builder.dart";
 export "src/widgets/generic_widgets/searchable_indexed_list_view_builder.dart";
 export "src/widgets/helpers/maybe_widget.dart";
-export "src/widgets/language/language_flag_picker.dart";
 export "src/widgets/language/language_picker.dart";
 export "src/widgets/language/language_tile.dart";
 export "src/widgets/phone_code/phone_code_picker.dart";
